Add files via upload
This was our first exploratory data analysis project on diamond sales.
We cleaned the data, explored it and made some basic linear regression models. 

We worked through excel files and used R Studio code for the processing of the data and 
analysis. The full report is in the PDF format